requirements.txt
# Core Streamlit framework
streamlit==1.28.1

# Document processing
python-docx==0.8.11

# Machine Learning and NLP - optimized for Streamlit Cloud
transformers==4.35.2
torch==2.1.1+cpu
sentence-transformers==2.2.2
tokenizers==0.15.0

# Core utilities
numpy==1.24.3
pandas==2.0.3

# Text processing
scikit-learn==1.3.2

# Additional dependencies for deployment stability
requests==2.31.0
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Hugging Face Hub for model downloads
huggingface-hub==0.19.4

# File handling and utilities
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0

# Security and compatibility
packaging==23.2
typing-extensions==4.8.0

# Memory monitoring and optimization
psutil==5.9.6